HorizontalBarChart: migrated enzyme tests to react tests
